# Optimize GetCommitHistory