--- a/src/docfx/config/AppData.cs
+++ b/src/docfx/config/AppData.cs
@@ -48,7 +48,7 @@ namespace Microsoft.Docs.Build
 
         public static string GetCommitCachePath(string remote)
         {
-            return Path.Combine(CacheRoot, "commits", HashUtility.GetMd5Hash(remote));
+            return Path.Combine(CacheRoot, "commits2", HashUtility.GetMd5Hash(remote));
         }
 
         public static string GetCommitBuildTimePath(string remote, string branch)