--- a/src/docfx/lib/git/GitCommitProvider.cs
+++ b/src/docfx/lib/git/GitCommitProvider.cs
@@ -39,7 +39,7 @@ namespace Microsoft.Docs.Build
         {
             foreach (var p in _fileCommitProvidersByRepoPath.Values)
             {
-                p.SaveCache();
+                p.Save();
             }
         }