… src/docfx/lib/git/GitCommit.cs
… src/docfx/lib/git/GitCommitCache.cs
… src/docfx/lib/git/LibGit2.cs
… src/docfx/lib/process/ProcessUtility.cs
… src/stylecop.ruleset