Optimize `GetCommitHistory`:

- Eliminate `LoadCommit` call when commit cache present.
- Pool objects to avoid excessive memory usage when commit cache missing

legend  | peak memory usage without cache    | build time without cache | build time with cache | cache size
---|---|---|---|---
Before | 7.8G | 5:33 | 1:13 | 3MB
After | 7G | 5:50 | 1:03 | 6MB


https://dev.azure.com/ceapex/Engineering/_workitems/edit/113226/

###### Microsoft Reviewers: [Open in CodeFlow](http://wpcp.azurewebsites.net/CodeFlowProtocolProxy2.php?pullrequest=https://github.com/dotnet/docfx/pull/5027)